@@ -1,3 +1,3 @@
 module github.com/rhnvrm/simples3
 
-go 1.13
+go 1.25.0
